Fix parts of label/move/clear
So we have these special labels like \Inbox, which represent INBOX or \Draft
which represents gmail.Drafts. I have these mapped so that now if I use a label
of INBOX, it will change it to what I meant \Inbox and vice-versa if I used
\Inbox in move. I've had some rules I've made that were originally labels, but
then became moves and vice-versa and I forgot to make this change. Hilarity
ensued.