this is a dumb fix, but i can tolerate it for now
The problem is that font width seems to work weird on Mac. This could be
a bug in zsh or in kitty, I haven't really worked it out. Anyway,
basically, when I get a fresh prompt after hitting enter, I get one
space between teh prompt and cursor. However, if I go away and come
back, the terminal gets redrawn with two spaces between the prompt and
the command that was typod on the line.

This does not happen on Ubuntu. This did not happy on Mac with the
previous clock emoji. The best fix will probably be to find a
single-width clock emoji like the one I was using before that works on
both Mac and Linux. The one I was using on Mac did not work on Linux.
However, I can probably fix that by finding a font with the regular
clock emoji I was using on Mac on Linux and then using kitty's symbol
table config to remap that symbol into the Fira Code font on Linux.

Whee!

I write all this down here to help me remember when I come back to this
later. Maybe I will even go back through the logs and read it. Or maybe
I will fix it later today or tomorrow and won't need to.

Am I rambling? I feel like I'm rambling. I must really not want to do
the task I'm supposed to do at work, huh? Uyup. Don't want to do it. I
wonder what other productivity improvements I can make for myself
instead of working...?